fix resnet50 models to work with deeplift

+++ experiments/medical_imaging/lib/datasets/cbis_ddsm.py
@@ -81,7 +81,7 @@ class CBIS_DDSM_patches(Dataset):
         else:
             i = cv.imread(path, cv.IMREAD_ANYDEPTH)
             i = (i - i.min()) / (i.max() - i.min())
-            image = i
+            image = i.astype(np.float32)
 
         image = self.transforms(image)
         image = image.expand(3,*image.shape[1:])
